Move MeterReaderTask and MeterReaderNode into core module
